+++ src/BOproject/server/router/Router.java
@@ -5,12 +5,14 @@ import java.net.InetSocketAddress;
 
 
 
+
 import com.sun.net.httpserver.HttpServer;
 
+import BOproject.server.path.AiRecomendServer;
 import BOproject.server.path.ArticleDeleteServer;
+import BOproject.server.path.ArticleLikeCountServer;
 import BOproject.server.path.ArticleListServer2;
 import BOproject.server.path.ArticleRegisterServer;
-import BOproject.server.path.ArticleRegisterServer2;
 import BOproject.server.path.LoginCheckServer;
 import BOproject.server.path.ProductListServer;
 import BOproject.server.path.UserRegisterServer;
@@ -27,8 +29,10 @@ public class Router {
 		server.createContext("/register", new UserRegisterServer());
 		server.createContext("/productlist", new ProductListServer());
 		server.createContext("/articlelist", new ArticleListServer2());
-		server.createContext("/articleregist", new ArticleRegisterServer2());
+		server.createContext("/articleregist", new ArticleRegisterServer());
 		server.createContext("/articledelete",new ArticleDeleteServer());
+		server.createContext("/articlelikecount", new ArticleLikeCountServer());
+		server.createContext("/airecomend", new AiRecomendServer());
 		
 		server.setExecutor(null);
 		server.start();
